feat: add validity state for controls
Add a delay when restoring the connection.
Previously, when communication was restored, if there were changes on the "buttons" control devices, then the rule for switching on or off the relay immediately triggered.

Now the rule is triggered only if there is a connection with the device and the control change is ignored when the connection is restored. Added a new "valid" property. When the connection is restored, a delay of 2 seconds is triggered, and only then "valid" becomes true. When changing the "buttons" controls, the "valid" property is also checked.